requirements.txt
anyio==3.6.1
appdata==2.1.2
asgiref==3.5.2
asyncpg==0.25.0
bcrypt==3.2.2
cffi==1.15.0
click==8.0.3
colorama==0.4.5
databases==0.6.0
Django==4.0.5
fastapi==0.78.0
fastapi-pagination==0.9.3
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
image==1.5.33
passlib==1.7.4
Pillow==9.1.1
psycopg2==2.9.3
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
qrcode==7.3.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
sqlparse==0.4.2
starlette==0.19.1
to-requirements.txt==1.1.8
typing_extensions==4.2.0
tzdata==2022.1
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.3
